Extract Pod function result normalization from the callback route
Move the existing runner/legacy envelope parsers into a shared module with
no behavior change so later protocol work can land as a small additive diff.
